requirements-ha-stubs.txt
# Minimal Home Assistant pytest dependencies for integration stub tests.
homeassistant
pytest-homeassistant-custom-component
